Support for multiline matching
I feel this is kind of hackish, but it's better to be able to parse it
than to fail at it.

In an ideal situation we could remove the extra whitespace and validate
the config, but it would be very hard to do that with complex if
statements in a config.

My current goal is to be able to read anything and write valid
dhcpd.conf files.  A future goal might be to read anything and write the
same results without mangling, but then mangle if changes happen.