Added new Authoritative.pm module
This is a one-off because authoritative is a special case keyvalue that
usually has a blank value "authoritative ;" but can have no value
"authoritative;"

It was difficult to get KeyValue to accept this so I decided to make a
separate module.  If I find any others like it then It'll be easier to
handle it with this module than to hack KeyValue.

fixes #4