Change parser to load modules on demand
On each line of the config file, the parser used to check if child
modules matched the lines regex.  To do this it loaded all candidate
children modules then at each new line would check them all.

The problem with this is that the loading is largely unneeded and
resource intensive.  Any module like "include" that had a bunch of
possible children would end up loading all those modules even if the
include file was blank or never parsed.

This potentially fixes #6.