perl version bumped to 5.12.5
There was a change between 5.10 and 5.11 to the way filehandles work.
Lots of the code depends on being able to use getpos.  Rather than
figuring out what broke between those two old versions, I've set 5.12.5
as the minimum version.

That means this won't work on Centos v5, but it's been broken for that
platform probably from the beginning, now it is documented.