removing unneeded tests
I've copied a bunch of test files around and left tests in that were
duplicates.  I was doing that both because I was lazy but also because I
wanted to test things extensively under different circumstances,
including different config files that were similar but not the same.

Now that I'm running make test more often I don't want it to take
forever. :)