fixed children statement
was caused by precedence error.

Slurp and Includes are basically the only things not working now, but
the speed isn't great.  It's a little faster than the old parser, but
it's still an order of magnitude slower than I'd like it to be.

It's an intermediate stage right now though.  Most of the regexes need
to be cleaned up but I wanted to see if I could get it basically working
before fixing it.

Stuff like this:

        if ($arg2 eq ';') {
            $arg1 .= ';';
        }

Can go away once we rewrite the regexes.  And all of them can be
anchored with ^$ since we always take care of surrounding whitespaces.