fix no space after brace
This fixes #26 (with a small caveat)

Lines like this now work:

    host workstation00 {hardware ethernet A1:B2:C3:D4:E5:03;

but a line like this will probably break things (at least on
$config->generate)

    option host-name "hello{world";

I could see braces being used in password fields or other strings, so it
is possible this will come up.

If it does I'll try to write a workaround for it.  I'm starting to think
that nothing short of a full parser will do things right, but I don't
have time to write it.